+++ R/pad_raster.R
@@ -226,8 +226,9 @@ pad_raster.matrix <- function(landscape,
                      pad_raster_cells = pad_raster_cells,
                      global = global)
 
-    if (return_raster || to_disk) {
-        warning("return_raster = TRUE or to_disk = TRUE not able for matrix input")
+    if (return_raster) {
+        warning("'return_raster = TRUE' or 'to_disk = TRUE' not able for matrix input.",
+                call. = FALSE)
     }
 
     return(result)