+++ tests/testthat/test-list-lsm.R
@@ -38,7 +38,7 @@ test_that("simplify returns vector", {
 test_that("returns warning if what and other argument is specified", {
 
     expect_warning(list_lsm(level = "landscape", what = "class"),
-                   grep = "Only using 'what' argument.",
+                   regexp = "Only using 'what' argument.",
                    fixed = TRUE)
 
     result <- list_lsm(level = "landscape", what = "class", verbose = FALSE)
@@ -77,18 +77,15 @@ test_that("Negative subset works for list_lsm()", {
 test_that("list_lsm returns error", {
 
     expect_error(list_lsm(what = "-patch"),
-                 grep = "Negative strings not allowed for 'what' argument.
-                 Please use other arguments for negative subsets.",
+                 regexp = "Negative strings not allowed for 'what' argument. Please use other arguments for negative subsets.",
                  fixed = TRUE)
 
     expect_error(list_lsm(level = c("-patch", "landscape")),
-                 grep = "Mixing of positive and negative strings as subset not
-                 allowed for the same argument.",
+                 regexp = "Mixing of positive and negative strings as subset not allowed for the same argument.",
                  fixed = TRUE)
 
     expect_error(list_lsm(what = "lsm_p_made_up"),
-                 grep = "Selected metrics do not exist. Please use 'list_lsm()'
-                 to see all available metrics.",
+                 regexp = "Selected metrics do not exist. Please use 'list_lsm()' to see all available metrics.",
                  fixed = TRUE)
 
 })