+++ tests/testthat/test-lsm-l-rpr.R
@@ -3,10 +3,10 @@ context("landscape level lsm_l_rpr metric")
 landscapemetrics_landscape_landscape_value <- lsm_l_rpr(landscape, classes_max = 5)
 
 test_that("lsm_l_rpr is typestable", {
-    expect_is(lsm_l_rpr(landscape), "tbl_df")
-    expect_is(lsm_l_rpr(landscape_stack), "tbl_df")
-    expect_is(lsm_l_rpr(landscape_brick), "tbl_df")
-    expect_is(lsm_l_rpr(landscape_list), "tbl_df")
+    expect_is(lsm_l_rpr(landscape, classes_max = 5), "tbl_df")
+    expect_is(lsm_l_rpr(landscape_stack, classes_max = 5), "tbl_df")
+    expect_is(lsm_l_rpr(landscape_brick, classes_max = 5), "tbl_df")
+    expect_is(lsm_l_rpr(landscape_list, classes_max = 5), "tbl_df")
 })
 
 test_that("lsm_l_rpr returns the desired number of columns", {