+++ tests/testthat/test-lsm-l-te.R
@@ -25,7 +25,7 @@ test_that("lsm_l_te returns in every column the correct type", {
 test_that("lsm_l_te option count_boundary is working", {
     te_with_boundary <- lsm_l_te(landscape, count_boundary = TRUE)
     te_without_boundary <- lsm_l_te(landscape, count_boundary = FALSE)
-    expect_less_than(te_without_boundary$value, te_with_boundary$value)
+    expect_lt(te_without_boundary$value, te_with_boundary$value)
 })
 
 test_that("lsm_l_te can handle raster with different xy resolution", {