+++ tests/testthat/test-sample-lsm.R
@@ -16,6 +16,9 @@ poly_2 <- sp::Polygons(list(poly_2), "p2")
 
 sample_plots <- sp::SpatialPolygons(list(poly_1, poly_2))
 
+# wrong plots
+sample_points_wrong <- cbind(sample_points, 1)
+
 # use lines
 x1 <- c(1, 5, 15, 10)
 y1 <- c(1, 5, 15, 25)
@@ -23,14 +26,16 @@ y1 <- c(1, 5, 15, 25)
 x2 <- c(10, 25)
 y2 <- c(5, 5)
 
-sample_lines <- sp::SpatialLines(list(sp::Lines(list(sp::Line(cbind(x1, y1)), sp::Line(cbind(x2, y2))), ID = "a")))
+sample_lines <- sp::SpatialLines(list(sp::Lines(list(sp::Line(cbind(x1, y1)),
+                                                     sp::Line(cbind(x2, y2))), ID = "a")))
 
 test_that("sample_lsm works for a matrix", {
 
     result_mat <- sample_lsm(landscape,
-                             y = sample_points, size = 15,
+                             y = sample_points, size = 5,
                              shape = "circle",
-                             what = c("lsm_l_ta", "lsm_l_np"))
+                             what = c("lsm_l_ta", "lsm_l_np"),
+                             verbose = FALSE)
 
     expect_is(object = result_mat, class = "tbl_df")
     expect_true(all(c("np", "ta") %in% result_mat$metric))
@@ -39,9 +44,10 @@ test_that("sample_lsm works for a matrix", {
 test_that("sample_lsm works for sp points", {
 
     result_sp <- sample_lsm(landscape,
-                            y = points_sp, size = 15,
+                            y = points_sp, size = 5,
                             shape = "square",
-                            what = "lsm_l_np")
+                            what = "lsm_l_np",
+                            verbose = FALSE)
 
     expect_is(object = result_sp, class = "tbl_df")
 
@@ -51,19 +57,20 @@ test_that("sample_lsm works for sp points", {
 test_that("sample_lsm works for polygons ", {
 
     result_poly <- sample_lsm(landscape,
-                              y = sample_plots,
-                              level = "patch")
+                              y = sample_plots, size = 5,
+                              level = "patch",
+                              verbose = FALSE)
 
     expect_is(object = result_poly, class = "tbl_df")
 
     expect_true(all("patch" %in% result_poly$level))
 
     if (!nzchar(system.file(package = "rgeos"))) {
+
         expect_warning(sample_lsm(landscape,
-                                  y = sample_plots,
+                                  y = sample_plots, size = 5,
                                   what = "lsm_p_area"),
-                       grep = "Package 'rgeos' not installed. Please make sure
-                       polygons are disaggregated.",
+                       regexp = "Package 'rgeos' not installed. Please make sure polygons are disaggregated.",
                        fixed = TRUE)
     }
 })
@@ -76,8 +83,7 @@ test_that("sample_lsm works for lines ", {
                                 y = sample_lines,
                                 size = 5,
                                 level = "landscape"),
-                     grep = "To sample landscape metrics in buffers around lines,
-                     the package 'rgeos' must be installed.",
+                     regexp = "To sample landscape metrics in buffers around lines, the package 'rgeos' must be installed.",
                      fixed = TRUE)
 
     }
@@ -87,7 +93,8 @@ test_that("sample_lsm works for lines ", {
         result_lines <- sample_lsm(landscape,
                                    y = sample_lines,
                                    size = 5,
-                                   level = "landscape")
+                                   level = "landscape",
+                                   verbose = FALSE)
 
         expect_is(object = result_lines, class = "tbl_df")
 
@@ -95,30 +102,60 @@ test_that("sample_lsm works for lines ", {
     }
 })
 
-test_that("sample_lsm works forwards arguments to calculate_lsm", {
+test_that("sample_lsm forwards arguments to calculate_lsm", {
 
     result_mat <- sample_lsm(landscape,
-                             y = sample_points, size = 15,
+                             y = sample_points, size = 5,
                              shape = "circle",
                              what = "lsm_p_core",
-                             edge_depth = 100)
+                             edge_depth = 100,
+                             verbose = FALSE)
 
     expect_true(all(result_mat$value == 0))
+})
+
+test_that("sample_lsm uses sample_ids", {
+
+    result <- sample_lsm(landscape,
+                         y = sample_points,
+                         plot_id = c(5, 25, 15),
+                         size = 15,
+                         shape = "circle",
+                         what = "lsm_l_ta",
+                         verbose = FALSE)
+
+    expect_equal(result$plot_id, expected = c(5, 15, 25))
+
+    result_wrong_id <- sample_lsm(landscape,
+                                  y = sample_points,
+                                  plot_id = c(5, 25, 15, 1),
+                                  size = 15,
+                                  shape = "circle",
+                                  what = "lsm_l_ta",
+                                  verbose = FALSE)
+
+    expect_equal(result_wrong_id$plot_id, expected = 1:3)
 })
 
 test_that("sample_lsm works for all data type", {
 
     result_stack <- sample_lsm(landscape_stack,
                                y = sample_plots,
-                               what = "lsm_l_ta")
+                               size = 5,
+                               what = "lsm_l_ta",
+                               verbose = FALSE)
 
     result_brick <- sample_lsm(landscape_brick,
                               y = sample_plots,
-                              what = "lsm_l_ta")
+                              size = 5,
+                              what = "lsm_l_ta",
+                              verbose = FALSE)
 
     result_list <- sample_lsm(landscape_list,
                               y = sample_plots,
-                              what = "lsm_l_ta")
+                              size = 5,
+                              what = "lsm_l_ta",
+                              verbose = FALSE)
 
     expect_is(result_stack, class = "tbl_df")
     expect_is(result_brick, class = "tbl_df")
@@ -140,26 +177,38 @@ test_that("sample_lsm works for all data type", {
 test_that("sample_lsm returns errors", {
 
     expect_error(sample_lsm(landscape,
-                            y = sample_points, size = 15,
-                            shape = "rectangle",
-                            what = c("lsm_l_ta", "lsm_l_np")),
-                 grep = "Shape = rectangle unknown.", fixed = TRUE)
+                            y = 1:3,
+                            size = 5),
+                 regexp = "'y' must be a matrix, SpatialPoints, SpatialLines or SpatialPolygons.",
+                 fixed = TRUE)
 
     expect_error(sample_lsm(landscape,
-                            y = 1:3),
-                 grep = "'y' must be a matrix, SpatialPoints, SpatialLines
-                 or SpatialPolygons.",
+                            y = sample_points,
+                            size = c(5, 15),
+                            what = "lsm_l_ta"),
+                 regexp = "Please provide only one value as size argument.",
                  fixed = TRUE)
 })
 
 test_that("sample_lsm returns warnings", {
 
-    sample_points_wrong <- cbind(sample_points, 1)
+    expect_warning(sample_lsm(landscape,
+                              y = sample_points_wrong, size = 5,
+                              what = "lsm_l_pr"),
+                   regexp = "'y' should be a two column matrix including x- and y-coordinates.",
+                   fixed = TRUE)
 
     expect_warning(sample_lsm(landscape,
-                              y = sample_points_wrong, size = 15,
-                              what = "lsm_l_pr"),
-                   grep = "'y' should be a two column matrix including
-                   x- and y-coordinates.",
+                              y = sample_points, size = 50, what = "lsm_l_ta"),
+                   regexp = "Some of buffers extend over the landscape border. Consider decreasing the size argument value.",
+                   fixed = TRUE)
+
+    expect_warning(sample_lsm(landscape,
+                              y = sample_points,
+                              plot_id = c(5, 25, 15, 1),
+                              size = 15,
+                              shape = "circle",
+                              what = "lsm_l_ta"),
+                   regexp = "Length of plot_id is not identical to length of y. Using 1...n as plot_id.",
                    fixed = TRUE)
 })