+++ tests/testthat/test-show-lsm.R
@@ -1,18 +1,23 @@
 context("test-show-lsm")
 
 test_that("show_lsm returns a plot", {
+
     patches_area <- show_lsm(landscape, what = "lsm_p_area")
+
     expect_is(patches_area, "ggplot")
 })
 
 test_that("show_lsm returns a facet plot", {
+
     patches_area <- show_lsm(landscape, what = "lsm_p_area",
                              class = c(1, 3),
                              labels = FALSE)
+
     expect_is(patches_area$facet, "FacetWrap")
 })
 
 test_that("show_lsm can handle stacks, bricks and lists", {
+
     plots_list <- show_lsm(landscape_list, what = "lsm_p_area")
     plots_stack <- show_lsm(landscape_stack, what = "lsm_p_area")
     plots_brick <- show_lsm(landscape_brick, what = "lsm_p_area")
@@ -30,15 +35,14 @@ test_that("show_lsm can handle stacks, bricks and lists", {
 test_that("show_lsm returns warnings and errors", {
 
     expect_warning(show_lsm(landscape, what = "lsm_p_area", class = c(1, "global")),
-                   grep = "'global' and 'all' can't be combined with any other class-argument.",
+                   regexp = "'global' and 'all' can't be combined with any other class-argument.",
                    fixed = TRUE)
 
     expect_error(show_lsm(landscape, what = "lsm_p_invented_metric"),
-                 grep = "Please provide one patch level metric only.
-                 To list available metrics, run list_lsm(level = 'patch').",
+                 regexp = "Please provide one patch level metric only. To list available metrics, run list_lsm(level = 'patch').",
                  fixed = TRUE)
 
     expect_error(show_lsm(landscape, what = "lsm_p_area", class = 5),
-                 grep = "'class' must contain at least one value of a class existing in the landscape.",
+                 regexp = "'class' must contain at least one value of a class existing in the landscape.",
                  fixed = TRUE)
 })